requirements.txt
streamlit
google-genai
toml
gTTS
edge-tts
python-dotenv
pyinstaller
